Cleaned up documentation
Removed the docs and configuration for running the micro-services omniwitness. If any compelling use-cases for this arise then we could support this again, but it is far easier to run and reason about a single binary that has the configuration baked into it. If anyone wants to add their own log for testing then the omniwitness monolith can still be made to do that locally for testing.