Added cloudbuild docker config for trillian-opensource-ci
This is forked from the non-legacy version of the config. This is a short-term solution to build public docker images until we are able to complete the internal process to allow the transparency-dev project to expose artifacts to the public.

It's a bit naughty, but I also snuck in a small amount of tidy-up in
both files to delete a variable that wasn't used.