Move and tidy up the witness configuration (#37)
* Refactoring to move config to omniwitness

The omniwitness should not have been depending on the witness cmd internals. If we decide we don't want the witness cmd depending on the onmniwitness, then we need a third location for this to move to where we can have cmd and omniwitness point to that.

* Remove LogID from witness config

We're more opinionated on this now; Log ID is whatever the function says it is. Custom IDs will cause problems, so let's get rid of it now.

* Drop HashStrategy while we're tidying

It only supports one value and we can add it if/when we need it