Remove entry point to naked witness
There's no need to have a witness main file. We've already removed the main files for the feeders and distributors, so this achieves consistency. The argument in the docs for keeping this for developers of new feeders doesn't really hold water; devs can fork this repo and write new feeders as part of the omniwitness.