Fail on redirect changing request method (#190)
A PUT request that hits a redirect gets converted to a GET request to the new resource. All of the existing checks for 200 will pass, which makes it look like it distributed even though no mutation operation occurred. This is a massive trap that is way too easy to fall into.