Witness properly handles ratcheting forward from tree size 0 (#266)
* Added tests for witnesses updating from tree size 0

The tests show that this currently works, providing that the proof is empty. When the merkle library is updated to pull in transparency-dev/merkle#140, this will fail too. In both of the new test cases, the witness should ratchet forward when a new checkpoint is provided.

As a separate note, I recommend rewriting these tests at some near point in the future. They are currently quite brittle as they rely on hard coded test data with no obvious script to update them. An in-memory test log would be a great way to generate the test data in a flexible way.

* Support ratcheting forward from checkpoints for size 0

The only way to handle this is via a special case.