Basic load test for updating witnesses (#298)
Load test escalates until witness updates take longer than a deadline,
or a success criteria of updates/sec is reached. If the number of logs
allocated isn't sufficient, then it can be the case that the load test
will run forever as increasing the throttle doesn't create more load due
to updates for each log being performed consecutively.

The load test always creates deterministic logs, with the same keys and
values. On init, it speaks to the witness to set the log sizes to the
last witnessed sizes. This aids load testing as the witness state
doesn't need to be cleared before every run.

The log implementations are currently in-memory and store all hashes.
This can be replaced with a more svelte compact range version if needed,
but I backed out of this approach as generating the consistency proofs
was not trivial and I wanted to get the bulk of the code written before
getting stuck on optimizations. In practice, I haven't had any problems
from the memory of this hammer getting too big, so this is sufficient
for now.

When performing the load test, the config printed by the load test at
startup needs to be put in the logs.yaml file locally when building the
witness.