Make incorrect size error more descriptive (#389)
Fixes a bug in the Spanner persistence implementation where underlying error types were being lost rather then wrapped.

This caused higher levels to be confused about what course of action to take in order to rectify the issue.
Added a test to catch this if it happens again, and augmented the particular error message we were seeing to provide a bit more info.