Hide configuration implementation (#399)
Shift the configuration to be done mostly in terms of config.Log structs rather than a number of similar types in various packages, and hide the implementation of the configuration delivery mechanism from the things which ultimately interrogate/use that config.

This opens a path to allowing binaries using the library to store their configuration in whichever way makes most sense for them.