OmniGCP uses spanner for canonical log list (#402)
Introduces a Spanner-backed LogConfig implementation, and uses that to configure the OmniGCP witness.

In order to preserve the desirable feature that all omniwitness instances, by default, support the same checked-in list of known logs, OmniGCP uses the embedded version of that list to insert/update the log configuration in Spanner.

Since the canonical log config is distinct from the embedded list, the operator can additionally provision their own set of logs into the Spanner table in order to augment the list of supported logs.