Improve loadtest tool (#413)
Makes cmd/loadtest behave more like real logs would.

Previously, while the throttle implementation did send the correct number of requests per unit interval, it did so by releasing all tokens for the interval as quickly as they could be consumed. This meant that the DUT gets hit with qps concurrent update requests at the beginning of the interval, followed by no further traffic until the interval has passed.

With this change, tokens are released at a steady rate through the interval, which is likely to be more representative of real traffic.

This change also adds a max_qps flag, which, if set, caps the rate at which requests will be sent to the DUT.